refactor: split the terminal view, and say where its boundaries are
TerminalView had grown to 927 lines, past the 800 the repo holds itself to,
and most of what it held was not the terminal's lifecycle:

- the find box is its own component (TerminalSearchBar), which owns nothing:
  the query, the match count and every jump stay with the terminal, where
  the xterm addon lives;
- the file drop is its own hook (useTerminalDrop) — the depth counter, the
  four handlers and the file-drag test. The handlers are written onto the
  container one by one rather than spread, because a11y lint reads JSX
  attributes and a spread would hide the static-element-interaction the
  container answers for in its own comment;
- the xterm privates, the grid fit and the font load are lib/terminal/
  term-view.ts, and that file joins term-transport and term-perf in the
  coverage exclusions: every one of them needs a real Terminal or a real
  document. The arithmetic under the fit stays in term-fit.ts, covered;
- `decodeBase64` moves to term-frame.ts, the codec beside it. It is the same
  PTY payload arriving through the /events fallback instead of the socket,
  it is pure, and it now has the tests it could not have inside a component:
  a multi-byte sequence, a high byte a text decode would replace, an escape.

The view is 745 lines and reads as the lifecycle it is.

Two comments were also promising the wrong thing. `reviewSlots` said it
answered "in document order" and does not — sorting is buildSlots' job,
because a RangeSet is what has to be built in order — and codemirror-threads
now says why it reads low in the coverage report: everything past buildSlots
needs the widget's element, which only exists once CodeMirror calls toDOM.
