+++ frontend/src/lib/terminal/term-frame.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "vitest"
-import { decodeFrame, encodeFrame } from "./term-frame"
+import { decodeBase64, decodeFrame, encodeFrame } from "./term-frame"
 
 describe("term-frame", () => {
   it("round-trips id and payload", () => {
@@ -32,5 +32,30 @@ describe("term-frame", () => {
     expect(decodeFrame(new Uint8Array(0))).toBeNull()
     expect(decodeFrame(new Uint8Array([0, 65]))).toBeNull()
     expect(decodeFrame(new Uint8Array([10, 65]))).toBeNull()
+  })
+})
+
+// The /events fallback's payload. Every byte has to come back exactly as the PTY
+// wrote it: the encoding exists because the JSON envelope cannot carry raw bytes,
+// so a decode that went through text would break the sequences it protects.
+describe("decodeBase64", () => {
+  it("answers the empty payload for an empty string", () => {
+    expect(decodeBase64("")).toEqual(new Uint8Array())
+  })
+
+  it("round-trips a multi-byte UTF-8 sequence byte for byte", () => {
+    // "ok ✓" — the tick is three bytes.
+    expect([...decodeBase64("b2sg4pyT")]).toEqual([0x6f, 0x6b, 0x20, 0xe2, 0x9c, 0x93])
+  })
+
+  it("keeps a high byte a text decode would replace", () => {
+    // 0xff is not valid UTF-8 on its own; the tail of a sequence split across
+    // two frames arrives exactly like this.
+    expect([...decodeBase64("/w==")]).toEqual([0xff])
+  })
+
+  it("keeps an escape sequence intact", () => {
+    // ESC [ 2 m
+    expect([...decodeBase64("G1sybQ==")]).toEqual([0x1b, 0x5b, 0x32, 0x6d])
   })
 })
